infra/pyproject.toml
[project]
name = "infra"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.234.1,<3.0.0",
    "constructs>=10.0.0,<11.0.0",
]

[dependency-groups]
dev = [
    "hypothesis>=6.150.3",
    "pytest>=8.4.2",
]
local = [
    "localstack>=3.0.0",
    "awscli-local>=0.22",
    "boto3>=1.34.0",
]
